Extract runtime options support implementation
